+++ Source/Main Dialogs/HardwareWizard/SubviewTableViewController.m
@@ -121,7 +121,7 @@
     }
 }
 
-- (BOOL) tableView:(NSTableView *) tableView shouldEditTableColumn:(NSTableColumn *) tableColumn row:(int) row
+- (BOOL) tableView:(NSTableView *) tableView shouldEditTableColumn:(NSTableColumn *) tableColumn row:(NSInteger) row
 {
     if ([self isValidDelegateForSelector: _cmd]){
 	return [[self delegate] tableView: tableView shouldEditTableColumn: tableColumn row: row];
@@ -131,7 +131,7 @@
     }
 }
 
-- (BOOL) tableView:(NSTableView *) tableView shouldSelectRow:(int) row
+- (BOOL) tableView:(NSTableView *) tableView shouldSelectRow:(NSInteger) row
 {
     if ([self isValidDelegateForSelector: _cmd]){
 	return [[self delegate] tableView: tableView shouldSelectRow: row];
@@ -151,7 +151,7 @@
     }
 }
 
-- (void) tableView:(NSTableView *) tableView willDisplayCell:(id) cell forTableColumn:(NSTableColumn *) tableColumn row:(int) row
+- (void) tableView:(NSTableView *) tableView willDisplayCell:(id) cell forTableColumn:(NSTableColumn *) tableColumn row:(NSInteger) row
 {
     if (tableColumn == subviewTableColumn){
         // This is one of the few interesting things going on in this class. This is where
@@ -199,9 +199,9 @@
 
 // Methods from NSTableDataSource protocol
 
-- (int) numberOfRowsInTableView:(NSTableView *) tableView
+- (NSInteger) numberOfRowsInTableView:(NSTableView *) tableView
 {
-    int count = 0;
+    NSInteger count = 0;
     
     if ([self isValidDelegateForSelector: _cmd]){
 	count = [[self delegate] numberOfRowsInTableView: tableView];
@@ -210,7 +210,7 @@
     return count;
 }
 
-- (BOOL) tableView:(NSTableView *) tableView acceptDrop:(id <NSDraggingInfo>) info row:(int) row dropOperation:(NSTableViewDropOperation) operation
+- (BOOL) tableView:(NSTableView *) tableView acceptDrop:(id <NSDraggingInfo>) info row:(NSInteger) row dropOperation:(NSTableViewDropOperation) operation
 {
     if ([self isValidDelegateForSelector: _cmd]){
 	return [[self delegate] tableView: tableView acceptDrop: info row: row dropOperation: operation];
@@ -220,7 +220,7 @@
     }
 }
 
-- (id) tableView:(NSTableView *) tableView objectValueForTableColumn:(NSTableColumn *) tableColumn row:(int) row
+- (id) tableView:(NSTableView *) tableView objectValueForTableColumn:(NSTableColumn *) tableColumn row:(NSInteger) row
 {
     id obj = nil;
 
@@ -231,14 +231,14 @@
     return obj;
 }
 
-- (void) tableView:(NSTableView *) tableView setObjectValue:(id) obj forTableColumn:(NSTableColumn *) tableColumn row:(int) row
+- (void) tableView:(NSTableView *) tableView setObjectValue:(id) obj forTableColumn:(NSTableColumn *) tableColumn row:(NSInteger) row
 {
     if ((tableColumn != subviewTableColumn) && [self isValidDelegateForSelector: _cmd]){
 	[[self delegate] tableView: tableView objectValueForTableColumn: tableColumn row: row];
     }
 }
 
-- (NSDragOperation) tableView:(NSTableView *) tableView validateDrop:(id <NSDraggingInfo>) info proposedRow:(int) row proposedDropOperation:(NSTableViewDropOperation) operation
+- (NSDragOperation) tableView:(NSTableView *) tableView validateDrop:(id <NSDraggingInfo>) info proposedRow:(NSInteger) row proposedDropOperation:(NSTableViewDropOperation) operation
 {
     if ([self isValidDelegateForSelector: _cmd]){
 	return [[self delegate] tableView: tableView validateDrop: info proposedRow: row proposedDropOperation: operation];
@@ -251,10 +251,10 @@
 - (BOOL) tableView:(NSTableView *) tableView writeRows:(NSArray *) rows toPasteboard:(NSPasteboard *) pboard
 {
     if ([self isValidDelegateForSelector: _cmd]){
-	return [[self delegate] tableView: tableView writeRows: rows toPasteboard: pboard];
+        return [[self delegate] tableView: tableView writeRows: rows toPasteboard: pboard];
     }
     else {
-	return NO;
+        return NO;
     }
 }
