+++ Source/Objects/Hardware/PXI/Pxi Base Classes/ORPxiBusProtocol.h
@@ -23,34 +23,34 @@
 - (void) resetContrl;
 - (void) checkStatusErrors;
 
-- (void) readLongBlock:(unsigned long *) readAddress
-			 atAddress:(unsigned int) pxiAddress
-			 numToRead:(unsigned int) numberLongs;
+- (void) readLongBlock:(uint32_t *) readAddress
+			 atAddress:(uint32_t) pxiAddress
+			 numToRead:(uint32_t) numberLongs;
 
-- (void) readLong:(unsigned long *) readAddress
-		atAddress:(unsigned long) pxiAddress
-	  timesToRead:(unsigned int) numberLongs;
+- (void) readLong:(uint32_t *) readAddress
+		atAddress:(uint32_t) pxiAddress
+	  timesToRead:(uint32_t) numberLongs;
 
 
-- (void) writeLongBlock:(unsigned long *) writeAddress
-			  atAddress:(unsigned int) pxiAddress
-			 numToWrite:(unsigned int) numberLongs;
+- (void) writeLongBlock:(uint32_t *) writeAddress
+			  atAddress:(uint32_t) pxiAddress
+			 numToWrite:(uint32_t) numberLongs;
 
 - (void) readByteBlock:(unsigned char *) readAddress
-			 atAddress:(unsigned int) pxiAddress
-			 numToRead:(unsigned int) numberBytes;
+			 atAddress:(uint32_t) pxiAddress
+			 numToRead:(uint32_t) numberBytes;
 
 - (void) writeByteBlock:(unsigned char *) writeAddress
-			  atAddress:(unsigned int) pxiAddress
-			 numToWrite:(unsigned int) numberBytes;
+			  atAddress:(uint32_t) pxiAddress
+			 numToWrite:(uint32_t) numberBytes;
 
 
 - (void) readWordBlock:(unsigned short *) readAddress
-			 atAddress:(unsigned int) pxiAddress
-			 numToRead:(unsigned int) numberWords;
+			 atAddress:(uint32_t) pxiAddress
+			 numToRead:(uint32_t) numberWords;
 
 - (void) writeWordBlock:(unsigned short *) writeAddress
-			  atAddress:(unsigned int) pxiAddress
-			 numToWrite:(unsigned int) numberWords;
+			  atAddress:(uint32_t) pxiAddress
+			 numToWrite:(uint32_t) numberWords;
 
 @end
